+++ src/BecklynRadBundle.php
@@ -3,6 +3,9 @@
 namespace Becklyn\RadBundle;
 
 use Becklyn\RadBundle\Bundle\BundleExtension;
+use Becklyn\RadBundle\DependencyInjection\DoctrineExtensionsCompilerPass;
+use Becklyn\RadBundle\Usages\EntityUsagesProviderInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
 /**
@@ -16,5 +19,17 @@ class BecklynRadBundle extends Bundle
     public function getContainerExtension ()
     {
         return new BundleExtension($this);
+    }
+
+
+    /**
+     */
+    public function build (ContainerBuilder $container) : void
+    {
+        $container->addCompilerPass(new DoctrineExtensionsCompilerPass());
+
+        $container
+            ->registerForAutoconfiguration(EntityUsagesProviderInterface::class)
+            ->addTag("entity_usages.provider");
     }
 }