+++ src/Route/DeferredRoute.php
@@ -1,0 +1,94 @@
+<?php declare(strict_types=1);
+
+namespace Becklyn\RadBundle\Route;
+
+use Symfony\Component\Routing\Exception\InvalidParameterException;
+use Symfony\Component\Routing\Exception\MissingMandatoryParametersException;
+use Symfony\Component\Routing\Exception\RouteNotFoundException;
+use Symfony\Component\Routing\Generator\UrlGeneratorInterface;
+
+/**
+ * A deferred route.
+ *
+ * Basically a VO encapsulating all of the logic required for generating a route, but without directly generating it.
+ * Instead the generation can be deferred to a later point in time (this way you don't need a dependency on the router
+ * right away).
+ */
+class DeferredRoute
+{
+    /**
+     * @var string
+     */
+    private $route;
+
+
+    /**
+     * @var array
+     */
+    private $parameters;
+
+
+    /**
+     * @var int
+     */
+    private $referenceType;
+
+
+    /**
+     * @param string $route         The route name.  #Route
+     * @param array  $parameters    the parameters required for generating the route
+     * @param int    $referenceType The reference type to generate for this route
+     */
+    public function __construct (string $route, array $parameters = [], int $referenceType = UrlGeneratorInterface::ABSOLUTE_PATH)
+    {
+        $this->route = $route;
+        $this->parameters = $parameters;
+        $this->referenceType = $referenceType;
+    }
+
+
+    /**
+     */
+    public function getRoute () : string
+    {
+        return $this->route;
+    }
+
+
+    /**
+     */
+    public function getParameters () : array
+    {
+        return $this->parameters;
+    }
+
+
+    /**
+     */
+    public function getReferenceType () : int
+    {
+        return $this->referenceType;
+    }
+
+
+    /**
+     * @throws RouteNotFoundException
+     * @throws MissingMandatoryParametersException
+     * @throws InvalidParameterException
+     */
+    public function generate (UrlGeneratorInterface $urlGenerator) : string
+    {
+        return $urlGenerator->generate($this->route, $this->parameters, $this->referenceType);
+    }
+
+
+    /**
+     * @return static
+     */
+    public function withParameters (array $parameters) : self
+    {
+        $modified = clone $this;
+        $modified->parameters = \array_replace($modified->parameters, $parameters);
+        return $modified;
+    }
+}